Fix TracksolidTrackService speed casting from (int) to (float) so exact decimals like 5.19 km/h are preserved

+++ idle-monitor/app/Services/TracksolidTrackService.php
@@ -69,7 +69,7 @@ class TracksolidTrackService
                     'guid'             => Str::uuid()->toString(),
                     'latitude'         => $track['lat'] ?? null,
                     'longitude'        => $track['lng'] ?? null,
-                    'speed'            => isset($track['gpsSpeed']) ? (int) $track['gpsSpeed'] : 0,
+                    'speed'            => isset($track['gpsSpeed']) ? (float) $track['gpsSpeed'] : 0,
                     'direction'        => isset($track['direction']) ? (int) $track['direction'] : 0,
                     'satellites'       => isset($track['satellite']) ? (int) $track['satellite'] : 0,
                     'acc_state'        => (isset($track['accStatus']) && $track['accStatus'] === 'ON') ? 1 : 0,
